+++ src/OptionalProp.php
@@ -2,11 +2,9 @@
 
 namespace Inertia;
 
-use Illuminate\Support\Facades\App;
-
 class OptionalProp implements IgnoreFirstLoad, Onceable
 {
-    use ResolvesOnce;
+    use ResolvesCallables, ResolvesOnce;
 
     /**
      * The callback to resolve the property.
@@ -34,6 +32,6 @@ class OptionalProp implements IgnoreFirstLoad, Onceable
      */
     public function __invoke()
     {
-        return App::call($this->callback);
+        return $this->resolveCallable($this->callback);
     }
 }