+++ src/ScrollProp.php
@@ -3,7 +3,6 @@
 namespace Inertia;
 
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\App;
 use Inertia\Support\Header;
 
 /**
@@ -14,9 +13,9 @@ use Inertia\Support\Header;
  *
  * @template T
  */
-class ScrollProp implements Mergeable
+class ScrollProp implements Deferrable, Mergeable
 {
-    use MergesProps;
+    use DefersProps, MergesProps, ResolvesCallables;
 
     /**
      * The property value.
@@ -125,6 +124,6 @@ class ScrollProp implements Mergeable
             return $this->resolved;
         }
 
-        return $this->resolved = is_callable($this->value) ? App::call($this->value) : $this->value;
+        return $this->resolved = $this->resolveCallable($this->value);
     }
 }